Added new role gcp to support contrail vgw
Contrail simple gateway (vgw) is used to provide external network connectivity
without any physical router, in test and development environments.

so gcp1 will be act as an edge router for virtual cloud/vagrant environment.

Since vgw cannot advertise its routes to any other system, one have to add route
manually to reach floating IPs.

In order to access default floating IP pool (100.1.0.0/24) from the machine in
side cloud network (say haproxy1), one have to add below route in it.

route add -net 100.1.0.0/24  gw 10.0.0.10 dev eth0

where 10.0.0.10 is gcp1 IP address, eth0 is the machine's interface by which it
connect to the gw, 100.1.0.0/24 is the floating IP network.

I am working on developing rjil::tempest, in which the code to add appropriate
route on tempest machine will be added as it is required for tempest to pass
floating IP tests.

Note:

* There is a race condition between floating IP pool and vgw
 creation. Created a consul kv after creating floating ip pool and blocked vgw
 creation util that kv available.

SPEC tests to be added on which I am working on.
This PR depends on PR #23 on jiocloud_contrail