Casandra and zookeeper should be up before contrail packages
Created a service blockers for some packages which start services to use the
cassandra and zookeeper, so that they will not be installed unless cassandra and
zookeeper are started. The packae installation can start the application and the
default configuration may be enough to write the data to either of these
application.

We have seen a possible situation when zookeeper is up, and contrail-api is
starting but cassandra is not writable there, in which case, contrail-api can
create some zookeeper entries while creating cassandra entries, but since
cassandra is not writable, that operation fail (writing cassandra). So whenever
they try this operation again, the code try to create the uids using zk, but
since zk have those object already exists, the operation fail and thus contrail
api fail (this must be a bug in contrail api though).

Please see https://github.com/JioCloud/puppet-rjil/issues/521 for more details